+++ packages/frontend/src/components/LoginProviderForm.vue
@@ -6,7 +6,7 @@
 
     <v-card-text>
       <v-list>
-        <v-form v-for="oidcProvider in apiSettings.oidcAuthProviders" :key="oidcProvider.id" @submit.prevent="auth.authProviderLoginBegin(oidcProvider, { reauth: props.reauth })">
+        <v-form v-for="oidcProvider in oidcAuthProviders" :key="oidcProvider.id" @submit.prevent="auth.authProviderLoginBegin(oidcProvider, { reauth: props.reauth })">
           <v-list-item>
             <s-btn-primary :key="oidcProvider.id" type="submit" block>
               Login with {{ oidcProvider.name }}
@@ -14,15 +14,15 @@
           </v-list-item>
         </v-form>
 
-        <slot v-if="apiSettings.remoteUserAuthProvider" name="remote">
+        <slot v-if="remoteUserAuthProvider" name="remote">
           <v-list-item>
-            <s-btn-primary @click="auth.authProviderLoginBegin(apiSettings.remoteUserAuthProvider, { reauth: props.reauth })" block>
-              Login with {{ apiSettings.remoteUserAuthProvider.name }}
+            <s-btn-primary @click="auth.authProviderLoginBegin(remoteUserAuthProvider, { reauth: props.reauth })" block>
+              Login with {{ remoteUserAuthProvider.name }}
             </s-btn-primary>
           </v-list-item>
         </slot>
 
-        <slot v-if="apiSettings.isLocalUserAuthEnabled" name="local">
+        <slot v-if="isLocalUserAuthEnabled" name="local">
           <v-list-item>
             <s-btn-secondary to="/login/local" block>
               Login with local user
@@ -43,6 +43,12 @@ const apiSettings = useApiSettings();
 const auth = useAuth();
 
 const props = defineProps<{
+  authProviders?: AuthProvider[];
   reauth?: boolean;
 }>();
+
+const authProviders = computed(() => props.authProviders || apiSettings.settings?.auth_providers || []);
+const oidcAuthProviders = computed(() => authProviders.value.filter(p => p.type === AuthProviderType.OIDC));
+const remoteUserAuthProvider = computed(() => authProviders.value.find(p => p.type === AuthProviderType.REMOTEUSER));
+const isLocalUserAuthEnabled = computed(() => authProviders.value.some(p => p.type === AuthProviderType.LOCAL));
 </script>